moved App component to its own component folder
