--- react-fundamentals/src/App.js
@@ -1,34 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-
-
-const Intro = (props) => (
-  <p className='App-intro'>
-    First functional component
-  </p>
-)
-
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-        <Intro/>
-      </header>
-    </div>
-  );
-}
-
-export default App;
