+++ react-fundamentals/src/components/Intro/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+
+const Intro = (props) => (
+    <p className='App-intro'>
+      {props.message}
+    </p>
+)
+
+
+export default Intro;
